int64_t

+++ src/blt/parse/argparse.cpp
@@ -527,7 +527,7 @@ namespace blt
             {
                 const auto& name = arg->a_flags.name;
                 std::cout << name;
-                auto size = std::max(static_cast<int64_t>(max_length) - static_cast<int64_t>(name.size()), 0l);
+                auto size = std::max(static_cast<int64_t>(max_length) - static_cast<int64_t>(name.size()), static_cast<int64_t>(0));
                 size += tab_size;
                 for (int64_t i = 0; i < size; i++)
                     std::cout << " ";
